Rename and move APIUtils class
